+++ domains/certificates/Query.API/API/GranularCertificateIssuer/EnergyMeasuredConsumer.cs
@@ -1,5 +1,7 @@
 using System.Threading.Tasks;
 using API.ContractService;
+using API.GranularCertificateIssuer.Repositories;
+using CertificateEvents.Aggregates;
 using CertificateEvents.Primitives;
 using MassTransit;
 using MeasurementEvents;
