+++ domains/certificates/Query.API/API/GranularCertificateIssuer/Repositories/ProductionCertificateRepository.cs
@@ -1,9 +1,10 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using CertificateEvents.Aggregates;
 using Marten;
 
-namespace API.GranularCertificateIssuer;
+namespace API.GranularCertificateIssuer.Repositories;
 
 public class ProductionCertificateRepository : IProductionCertificateRepository
 {
